requirements.txt
pyshark>=0.6
typer>=0.9.0
rich>=13.0.0
jinja2>=3.1.0
pydantic>=2.0.0
plotly>=5.18.0
tqdm>=4.66.0
colorama>=0.4.6
pyyaml>=6.0
kaleido>=0.2.1
requests>=2.31.0
openai>=1.0.0
anthropic>=0.7.0
